pokemon_data: move trade array to struct
This should help make modifications to the data easier from the
Flipper interface.

Note that right now this is kind of a hack, the originally named
DATA_BLOCK still exists as a pointer to the struct since right
now trade.cpp just expects a giant block of bytes.

Because of that, the struct format _needs_ to be packed. This
isn't ideal, but the compiler for the Flipper seems to be okay
with it and it does indeed work. Potentially, future iterations
could just use the struct itself and not need to cast to a byte
array.