trade: Add ability to reset connection state
Allows for resetting the connection back to not connected, as if
the Flipper left the Trade Center or restarted, but keeps the current
pokemon in RAM.

This woudl allow, for example, to trade a pokemon between two real
cartridges with only a single Game Boy. By trading from Pokemon Cart A
to the Flipper, resetting the connection state on the Flipper, and
loading Pokemon Cart B in to the Game Boy, and trading from the
Flipper to the real game.

Fixes #37

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>